deps: Update name of Flutter version to reflect new tag
Fixes today's instance of #1138, unbreaking CI.

The output of `flutter --version` at this commit changed because
the tag 3.28.0-0.0.pre was added (to a commit that's already been
in Flutter main since last week).

That was causing `tools/check flutter_version` to fail:
  https://chat.zulip.org/#narrow/channel/516-mobile-dev-help/topic/tools.2Fcheck.20flutter_version.20failure.20in.20CI.3F/near/1999743

Fix pubspec.yaml to use the new name of this version.