store [nfc]: Update outdated references to loadGlobalStore
These were supposed to be updated in commit
fc80be1a495a6c7ecd6769f865a318eb23bd3138
where we replaced loadGlobalStore with getGlobalStore.

`LiveGlobalBindings.load` is the only thing left analogous to the
previously referred to `loadGlobalStore` method. To avoid direct
referencce to this specific implementation on `GlobalStore`, mention
that loading it for the first time is expected to be asynchronous
instead.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>