macos: Refresh project files by jiggling a setting in Xcode GUI
Xcode made the project.pbxproj changes when I changed the "macOS
Deployment Target" setting in the GUI from 10.14 to 10.15 and back
again. The Runner.xcscheme changes happened when I built the app to
run.

A quick test confirmed that 7649477f4 explains the change from
"zulip.app" to "Zulip.app" (I tried a different string "asdfjkl" and
saw that it propagated in the same way).

Not sure what explains removing the "name = Pods;" line, but there's
a "path = Pods;" line right under it, and anyway the app continues
to build.