ci: Update workaround for broken ref for Flutter upstream main
Fixes #1177, take 2.

The original hacky workaround subsequently broke after the upstream
change that fixed the "upstream" vs. "origin" part of the issue.
We still need a version of this workaround, though, because we
clone the repo with the branch `main`, which in the Flutter repo
always points to the same commit as `master`.