compose: Skip failing test for windows.
The `attach from camera` test fails on
Windows due to the the path separator
used in the split function of the XFile.name
getter in `cross_file` being '\' instead of '/'.
This results in the getter returning the
entire path on Windows instead of the
last word present after the last '/' in the path.

Read this excellent explanation by
@chrisbobbe for better understanding:
https://chat.zulip.org/#narrow/channel/516-mobile-dev-help/topic/failing.20composebox.20test.20while.20running.20check.20script/near/2000301