deps: Upgrade share_plus{,_platform_interface} enabling more upgrades
Done with:

  flutter pub upgrade --major-versions share_plus{,_platform_interface}
(then commit, for the sake of a no-uncommitted-files check)
  tools/upgrade pub
(then squash)

See discussion:
  https://chat.zulip.org/#narrow/channel/243-mobile-team/topic/deps.3A.20running.20.60tools.2Fupgrade.20pub.60.20twice/near/2013483

All the Firebase packages are now upgraded to latest; none of them
appear in the `flutter pub outdated` list.

This seems to have unlocked upgrading `drift` too. Some followups
were needed:

- ran `tools/check drift --all --fix`

- ran `tools/check build_runner --all --fix`

- followed an instruction from the analyzer:

  info • 'package:drift_dev/api/migrations.dart' is deprecated and
    shouldn't be used. Import `package:drift/migrations_native.dart`
    instead • test/model/database_test.dart:4:1 •
    deprecated_member_use

Greg pointed out that it's awkward that some Drift files change, but
he found that they're not substantive changes, so it's fine:
  https://github.com/zulip/zulip-flutter/pull/1184#discussion_r1894471558
  https://github.com/zulip/zulip-flutter/pull/1184#discussion_r1894472486