home: Stop assuming account existence from loading page
We could pass realmUrl when initializing the `_LoadingPlaceholderPage`,
but that will still require a check for the existence of the account.

The loading page will contain a CircularLoadingIndicator when the
account does not exist. The user can't easily reach this page
because they can only logout from `ChooseAccountPage`, until we
start invalidating API keys. Even then, they will only see the
page briefly before getting navigated, so we should not show any
text at all.

Fixes: #1219

Signed-off-by: Zixuan James Li <zixuan@zulip.com>