db: Drop all tables on downgrade
We previously missed tables that are not known to the schema.  This
becomes an issue if a new table is added at a newer schema level. When
we go back to an earlier schema, that new table remains in the database,
so subsequent attempts to upgrade to the later schema level that adds
the table will fail, because it already exists.

Testing for this is blocked until #1172.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>