database test: Rewrite database tests using generated helpers
The tests are adapted from the test template generated from
`dart run drift_dev make-migrations`.

This saves us from writing the simple migration tests between schemas
without data in the future. For the test that upgrade the schema with
data, with the helper, while it ended up having more lines than the
original, the test becomes more structured this way.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>