actions test: Use a more realistic setup for logOutAccount
This rewrite ended being a bit more substantial than just switching both
accounts initial routes to the result of MessageListPage.buildRoute().

We do not use popUtil because the navigator stack normally should not
become empty, so the HomePage route for account1 stays. Additionally,
because we are pushing a different page route, we no longer need to set
up different messages as the discriminator, further simplifying the
test.

See also:
  https://github.com/zulip/zulip-flutter/pull/1183#discussion_r1902218275

Signed-off-by: Zixuan James Li <zixuan@zulip.com>