autocomplete: Simplify updating input on topic autocomplete
The point in this autocomplete UI is to choose the entire new value
for the topic input.  So we can do that directly, without the more
complicated logic that parallels what we need in the more
complicated case of autocomplete in the content input.

This is NFC as far as the text value of the input is concerned.  I
believe it's probably NFC entirely, because the remaining aspects of
TextEditingValue -- the selection and the composing region -- are
probably clobbered anyway by the next line's giving focus to the
content input.  (The new test already passes before this change.)