test [nfc]: Leave a few topics as explicit strings
These are being passed to test helpers which we'll leave as consuming
String instead of TopicName, for convenience in other test cases.

At these call sites the values were coming from fields that are or
will become TopicName, though; so when that happens and we make
TopicName no longer implicitly convertible to String, these will need
to be explicitly String instead.