deps: Upgrade firebase_core, firebase_messaging to latest
Upgrade carried using by following commands:
  flutter pub upgrade --major-versions firebase_messaging firebase_core
  pod update --project-directory=ios/
  pod update --project-directory=macos/

And adapted the test bindings to use the new
`providesAppNotificationSettings` flag.

Changelogs:
  https://pub.dev/packages/firebase_core/changelog#3100
  https://pub.dev/packages/firebase_messaging/changelog#1520

The one notable change there is a bump in the version of the
Firebase iOS SDK from 11.4.0 to 11.6.0 (seen also in the
Podfile.lock changes).

That changelog in turn is here:
  https://firebase.google.com/support/release-notes/ios

though it mixes in all the different Firebase components into one
changelog.  For the components we actually pull in, the changes all
look quite innocuous.

[greg: added readings of changelogs]