deps: Upgrade packages to latest, namely device_info_plus, mime, flutter_lints
This is the result of `tools/upgrade pub-major`.

Changelogs:
  https://pub.dev/packages/device_info_plus/changelog#1120
  https://pub.dev/packages/mime/changelog#200
  https://pub.dev/packages/flutter_lints/changelog#500

In device_info_plus, the one "breaking" change is:

    BREAKING FIX(device_info_plus): fixed webasm compliance (#3254).
    (e35e2123)

so I'll assume that's web-only and has no effect on us.

In mime, the breaking change is in a method "extensionFromMime",
which we don't use.

And flutter_lints by nature can't have broken anything if it doesn't
cause CI to fail.

There is one interesting change in flutter_lints: it removes three
lints that called for adding `const` modifiers.  Apparently (as
discussed in the dart-lang/lints thread linked from that changelog)
it's not clear that the performance benefits of doing that
everywhere was material.

[greg: added readings of changelogs]