action_sheet test: Cut assumption of topic recipient header in topic narrow
With #1039 "Fuse recipient header into app bar for topic narrows",
this setup will likely become invalid. It assumes that a topic
recipient header can be found in a topic-narrow message list.

Switching to the combined-feed narrow would break tests that use a
muted topic, because the topic's message(s) are excluded in that
narrow, so there wouldn't be a recipient header to tap.

Instead, trigger the topic action sheet from the topic-narrow app
bar.