do not merge; api: Indicate support for handling empty topics
Look for `allow_empty_topic_name` and `empty_topic_name` under "Feature
level 334" in the API Changelog to verify the affected routes:
  https://zulip.com/api/changelog

To keep the API bindings thin, instead of setting
`allow_empty_topic_name` for the callers, we require the callers to pass
the appropriate values instead.

Instead of making this parameter a `bool` that defaults to `false`
(and have the bindings remove the field when it's `false`), we type it
as `bool?` and only drop it when it is `null`.  This is also for making
the API binding thin.

Fixes: #1250