compose: Support sending to empty topic
This does not rely on TopicName.displayName being non-nullable or
"empty_topic_name" client capability, so it is not an NFC change.

The key change that allows sending to empty topic is that
`textNormalized` can now be actually empty, instead of being
converted to "(no topic)" with `_computeTextNormalized`.

---

This is accompanied with a content input hint text change, so that
"Message #stream > general chat" appears appropriately when we make
TopicName.displayName nullable.

---

Previously, "Message #stream > (no topic)" was the hint text for
content input as long as the topic input is empty and topics are not
mandatory.  Showing the default topic does not help create incentive
for the user to pick a topic first.  So only show it when they intend
to leave the topic empty.

See discussion:
  https://chat.zulip.org/#narrow/channel/530-mobile-design/topic/general.20chat.20design.20.23F1297/near/2088870

---

This does not aim to implement hint text changes to the topic input,
which is always "Topic".  We will handle that as a follow-up.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>