action_sheet test: Move some logic out of checkButtons helper
Upcoming tests for the resolve/unresolve button can use this to test
an edge case where the button isn't offered.

While we're at it, replace finder logic in some existing checks,
making some a bit more targeted (so a not-NFC change): when finding
`UserTopicUpdateButton`s by their label text, we now require the
text to be under a BottomSheet widget.