action_sheet test: Make some show-from-inbox setup optional in a `prepare`
To test showing the topic action sheet from the message list (app
bar or recipient header), we don't need `store.addMessage`; in those
cases, the message data comes via MessageListView's message fetch.

To test showing the action sheet from the Inbox, we do need a
message ID to be in the Unreads data. So, arrange that, but only in
the tests (one test) that use the Inbox.

While we're at it, also make this setup represent the common case
where we don't have the full Message object because a message list
containing the message hasn't been opened yet. When we write tests
for the upcoming resolve/unresolve button in the topic action sheet,
that's a case we'll want to exercise, in addition to the case where
we do have the message details.