doc: Add libdrm troubleshooting entry
I ran into this today when I tried running the Linux app.

(It looks like the reason I hadn't seen it before is that until
recently I had this package installed; it was installed
"automatically", i.e. only because it was a dependency of
something else, and then I upgraded my machine, it was no longer
such a dependency, and got autoremoved.  Now that I've installed it
directly, it'll stay.)