api [nfc]: Extract and parse UpdateMessageMoveData from UpdateMessageEvent
This data structure encapsulates some checks so that we can make all
fields non-nullable, with reasonable fallback values.  As of writing,
we do not use origStreamId (a.k.a.: 'stream_id') when there was no
message move, even though it is present if there were content edits
This makes dropping 'stream_id' when parsing `moveData` into `null`
acceptable for now.

Despite the NFC, a subtlety of this change is that `FormatException`'s
or null-check errors will be thrown earlier in the polling loop.
Because we now parse the move data upfront, instead of dealing with it
when handling the message event.  Both will end up resulting in the
event queue being replaced if such an error occurs.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>