action_sheet [nfc]: Comment on user-topic button about client-side check
Zixuan points out that this could be helpful here:
  https://github.com/zulip/zulip-flutter/pull/1301#discussion_r1932747085

This probably makes sense to not do until #1078, when we'll have
separate buttons for each of the four possible user-topic states.

This same comment *could* be added on various other action-sheet
buttons. But realistically it wouldn't make a difference on any that
we've implemented so far: all the others (besides this and
ResolveUnresolveButton) are on the message action sheet, and they're
all about actions that would only be done by the self-user --
starring, marking as unread, etc. -- so they can't be done without
the user's awareness, and will rarely be done anyway because you'd
need to do it on a different client.