sticky_header test: Favor drag gestures over taps, when they compete
This is nearly NFC.  The sense in which it isn't is that if any of
the `tester.drag` steps touched a header -- which listens for tap
gestures -- then this would ensure that step got interpreted as a
drag.  The old version would (a) interpret it as a tap, not a drag,
if it was less than 20px in length; (b) leave those first 20px out
of the effective length of the drag.

The use of DragStartBehavior.down fixes (b).  Then there are some
drags of 5px in these tests, subject to (a); TouchSlop fixes those
(by reducing the touch slop to 1px, instead of 20px).

This change will be needed in order to have the item widgets start
recording taps too, like the header widgets do, without messing up
the drags in these tests.