user: Split a UserStore out from PerAccountStore
Like ChannelStore and others, this helps reduce the amount of
complexity that's concentrated centrally in PerAccountStore.

This change is all NFC except that if we get a RealmUserUpdateEvent
for an unknown user, we'll now call notifyListeners, and so might
cause some widgets to rebuild, when previously we wouldn't.
That's pretty low-stakes, so I'm not bothering to wire through
the data flow to avoid it.