notif [nfc]: Note a todo-i18n for display names of notif sound files
Toward #277.

These may be tricky to actually handle translating, particularly
given the logic in _ensureInitNotificationSounds that looks at the
filenames of existing files.  They're also in a pretty obscure
corner where I expect very few users see them, so not a priority.

But they (a) are user-facing and (b) use English words ("chime",
"low", "high") to communicate, so are squarely in the category of
strings that should in principle get translated.  Mark them as such.