sticky_header [nfc]: Expand on the header-overflows-sliver case
This case has several bugs in it.  Not coincidentally, it's tricky to
think through: there are several sub-cases and variables involved
(the growth direction, vs. the header-placement direction, vs. the
coordinate direction, ...).  And in fact my original PR revision
which fixed the cases that would affect the Zulip message list
was still conceptually confused, as evidenced by the fact that it
turned out to break other cases:
  https://github.com/zulip/zulip-flutter/pull/1316#discussion_r1947339710

One step in sorting that out was the preceding commit which split
this overflows-sliver case from the alternative.  As a next step,
let's expand on the reasoning here a bit, with named variables
and comments.  In doing so, it becomes more apparent that several
points in this calculation are wrong; for this NFC commit, mark
those with TODO-comments.  We'll fix them shortly.