sticky_header [nfc]: Fix childMainAxisPosition to properly use paintExtent
This fixes a latent bug: this method would give wrong answers if the
sliver's paintExtent differed from its layoutExtent.

The bug is latent because performLayout currently always produces a
layoutExtent equal to paintExtent.  But we'll start making them differ
soon, as part of making hit-testing work correctly when a sticky
header is painted by one sliver but needs to encroach on the layout
area of another sliver.  The framework calls this method as part of
hit-testing, so that requires fixing this bug too.