sticky_header test [nfc]: Make "first/last item" finders more robust
The existing `.first` and `.last` versions rely on the order that
children appear in the Flutter element tree.

As is, that works because the `_Item` widgets are all children of one
list sliver, and it manages its children in a nice straightforward
order.  But when we add multiple list slivers, the order will depend
also on the order those appear as children of the viewport; and in
particular the items at the edges of the viewport will no longer
always be the first or last items in the tree.  So introduce a couple
of custom finders to keep finding the first or last items in the
sense we mean.

For examples of using the APIs these finders use, compare the
implementation of [FinderBase.first].