msglist test: Add and use helper checkAppBarChannelTopic
We'll use this for some new tests, coming up.

Also use it now for one test that's only been checking the topic in
the app bar. That test now checks the channel name too, as it was
before b1767b231, which dropped that specificity when we split the
app bar's channel name and topic onto two rows.