store: Expect AccountNotFoundException when reloading store
This can happen when `removeAccount` was called
during `doLoadPerAccount`, possibly due to the user logging out
from choose-account page while the store is getting reloaded.

However, it is currently not reachable live because of a bug in
`UpdateMachine.load`.  See #1354.

Modulo the bug, `UpdateMachine` can safely ignore this error
acknowledging that the reload has failed, given that the user
can continue using the app by navigating from choose-account page.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>