test [nfc]: Add the few missing awaits on handleEvent calls
This is NFC because there isn't currently actually anything to wait
for; the implementation of handleEvent is entirely synchronous.
But in principle it's async, because in the future it will be, for
the reasons described in baea6d746.

Thanks to Zixuan for spotting this pattern in some newly-added tests
(coming up):
  https://github.com/zulip/zulip-flutter/pull/1327#discussion_r1945527867