user [nfc]: Refer to "user store" rather than "store.users" in text
At the end of this series, there won't be a `users` member on the
store, so this code-like reference will no longer make sense.
The underlying meaning in each of these contexts will still be apt,
though: the data structure where the known users are tracked.  So
update them to make that reference more generically.