message-list : add a prompt to notify guest(s) when DM
Fixes #798
showDMWarningBanner could be initialized in the initState to shouldShowGuestUserWarningPrompt(store) and Visibility wrapper could have been gotten rid off , but due to unavailability of store beforehand it was not possible.

To create the guests list, I got rid off all the users in the store which are not in the recepients list and role-checked them.