msglist [nfc]: Use SliverPaintOrder.firstIsTop
This changes the paint order of these slivers.  Right now that has
no observable effect, because the second sliver has zero height
and paints nothing.

In the future when we put the messages in back-to-back slivers,
this is the paint order we'll need so that the top sliver can have
a sticky header overflow properly over the bottom sliver.